+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixing migrations issue that may arise due to `DataStore(initialize_db=True)` statement in `covalent start`
+- Fixed postprocessing workflows that return an electron with an incoming wait_for edge
 
 ## [0.173.0] - 2022-08-10